feat: add experimental updaters optimistic option
Experimental updaters modifyCache and modifyFragment access cache data
directly to inject data into modifiers, add ability to control which
cache that data originates from.

Default to using optimistic cache rather than normal cache: when no
optimistic responses have been used it is identical, and in the cases
where there is optimistic data in the cache we don't want to be unable
to access it in our updaters.