feat: experimental variants need cached query/frag
modifyFragment and modifyQuery currently use the entire cache to modify
their query or fragment, this makes it difficult to type and reason
about in terms of what values will be provided to field modifiers.

If we enforce that the passed query or fragment must exist in the cache
then it becomes possible to becomes possible to ensure that modifiers
are called with data in the correct schema. It also makes it more clear
when to use these updaters over writeFragment or writeQuery: when you
would need to readFragment or readQuery first you can instead use
modifyQuery or modifyFragment.